[Fix] Changed all value type of Int to Double. (for obvious reasons)

+++ CalculatorPlayground.playground/Contents.swift
@@ -1,16 +1,16 @@
 import Foundation
 
 protocol Operator {
-    var valueA: Int { get }
-    var valueB: Int { get }
+    var valueA: Double { get }
+    var valueB: Double { get }
     
-    func operate() -> Int
+    func operate() -> Double
 }
 
 class Calculator {
-    let valueA: Int
-    let valueB: Int
-    var result: Int?
+    let valueA: Double
+    let valueB: Double
+    var result: Double?
     
     // 호출되기 전엔 메모리를 차지하지 않게 lazy var 선언
     lazy var adder = Adder(valueA: valueA, valueB: valueB)
@@ -18,13 +18,13 @@ class Calculator {
     lazy var multiplier = Multiplier(valueA: valueA, valueB: valueB)
     lazy var divider = Divider(valueA: valueA, valueB: valueB)
     
-    init(valueA: Int, valueB: Int, result: Int? = nil) {
+    init(valueA: Double, valueB: Double, result: Double? = nil) {
         self.valueA = valueA
         self.valueB = valueB
         self.result = result
     }
     
-    func storeResult(of operation: () -> Int) {
+    func storeResult(of operation: () -> Double) {
         result = operation.self()
     }
 }
@@ -37,58 +37,58 @@ calculator.multiplier.operate()
 calculator.divider.operate()
 
 class Adder: Operator {
-    var valueA: Int
-    var valueB: Int
+    var valueA: Double
+    var valueB: Double
     
     
-    init(valueA: Int, valueB: Int) {
+    init(valueA: Double, valueB: Double) {
         self.valueA = valueA
         self.valueB = valueB
     }
     
-    func operate() -> Int {
+    func operate() -> Double {
         return valueA + valueB
     }
 }
 
 class Subtractor: Operator {
-    var valueA: Int
-    var valueB: Int
+    var valueA: Double
+    var valueB: Double
     
-    init(valueA: Int, valueB: Int) {
+    init(valueA: Double, valueB: Double) {
         self.valueA = valueA
         self.valueB = valueB
     }
     
-    func operate() -> Int {
+    func operate() -> Double {
         return valueA - valueB
     }
 }
 
 class Multiplier: Operator {
-    var valueA: Int
-    var valueB: Int
+    var valueA: Double
+    var valueB: Double
     
-    init(valueA: Int, valueB: Int) {
+    init(valueA: Double, valueB: Double) {
         self.valueA = valueA
         self.valueB = valueB
     }
     
-    func operate() -> Int {
+    func operate() -> Double {
         valueA * valueB
     }
 }
 
 class Divider: Operator {
-    var valueA: Int
-    var valueB: Int
+    var valueA: Double
+    var valueB: Double
     
-    init(valueA: Int, valueB: Int) {
+    init(valueA: Double, valueB: Double) {
         self.valueA = valueA
         self.valueB = valueB
     }
     
-    func operate() -> Int {
+    func operate() -> Double {
         valueA / valueB
     }
 }
